backend/requirements.txt
Django>=3.0,<4.0
psycopg2-binary
djangorestframework
djangorestframework-simplejwt
django-environ
djangorestframework-simplejwt[blacklist]
